Added city_name + prefix
thank you Bruno

+++ src/csv_writer.py
@@ -6,16 +6,16 @@ schema = ['ZONE_CODE', 'LAT', 'LON', 'CITY', 'DISPLAY_STREETNAME',
           'GOOGLE_STREETNAME', 'ZONE_STREET', 'ZONE_DESCRIPTION', 'Tariff_range', 'Unieke ID']
 
 
-def write_header(country_prefix, city_name):
-    file_name = f'data\\{country_prefix+city_name}.csv'
-    os.remove(file_name)
+def write_header(file_name):
+    if os.path.exists(file_name):
+        os.remove(file_name)
     with open(file_name, 'a', newline='') as file_name:
         writer = csv.writer(file_name)
         writer.writerow([header for header in schema])
 
 
-def write_rows(coordinate, city_name, country_prefix, street_name):
-    file_name = f'data\\{country_prefix+city_name}.csv'
+def write_rows(file_name,coordinate, city_name, country_prefix, street_name):
+    file_name = f'data\\{country_prefix}_{city_name}.csv'
 
     display_streetname = street_name
     google_streetname = display_streetname
